Minor snprintf() usage tweaks: 1) don't assume snprintf() returns -1 on error, check for <0 2) when comparing return value of sizeof(foo), cast the sizeof, not the len 3) cast return value to void in cases where snprintf cannot fail

+++ plugins/sudoers/find_path.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1996, 1998-2005, 2010-2015
+ * Copyright (c) 1996, 1998-2005, 2010-2015, 2017-2019
  *	Todd C. Miller <Todd.Miller@sudo.ws>
  *
  * Permission to use, copy, modify, and distribute this software for any
@@ -140,7 +140,7 @@ find_path(const char *infile, char **outfile, struct stat *sbp,
 	 */
 	len = snprintf(command, sizeof(command), "%.*s/%s",
 	    (int)(ep - cp), cp, infile);
-	if (len <= 0 || (size_t)len >= sizeof(command)) {
+	if (len <= 0 || len >= (int)sizeof(command)) {
 	    errno = ENAMETOOLONG;
 	    debug_return_int(NOT_FOUND_ERROR);
 	}
@@ -154,7 +154,7 @@ find_path(const char *infile, char **outfile, struct stat *sbp,
      */
     if (!found && checkdot) {
 	len = snprintf(command, sizeof(command), "./%s", infile);
-	if (len <= 0 || (size_t)len >= sizeof(command)) {
+	if (len <= 0 || len >= (int)sizeof(command)) {
 	    errno = ENAMETOOLONG;
 	    debug_return_int(NOT_FOUND_ERROR);
 	}
